+++ leptonica/src/main/java/org/bytedeco/javacpp/presets/lept.java
@@ -34,12 +34,15 @@ import org.bytedeco.javacpp.tools.InfoMapper;
     @Platform(include={"leptonica/alltypes.h", "leptonica/environ.h", "leptonica/array.h", "leptonica/bbuffer.h", "leptonica/heap.h", "leptonica/list.h",
         "leptonica/ptra.h", "leptonica/queue.h", "leptonica/stack.h", "leptonica/arrayaccess.h", "leptonica/bmf.h", "leptonica/ccbord.h",
         "leptonica/dewarp.h", "leptonica/gplot.h", "leptonica/imageio.h", "leptonica/jbclass.h", "leptonica/morph.h", "leptonica/pix.h",
-        "leptonica/recog.h", "leptonica/regutils.h", "leptonica/sudoku.h", "leptonica/watershed.h", "leptonica/allheaders.h"}, link="lept@.4"),
+        "leptonica/recog.h", "leptonica/regutils.h", "leptonica/sudoku.h", "leptonica/watershed.h", "leptonica/allheaders.h"}, link="lept@.4",
+        preloadpath="/usr/lib/", preload={"gif@.4", "jpeg@.62", "png16@.16", "tiff@.5", "webp@.5"}),
+    @Platform(value="linux-x86", preloadpath={"/usr/lib32/", "/usr/lib/"}),
+    @Platform(value="linux-x86_64", preloadpath={"/usr/lib64/", "/usr/lib/"}),
     @Platform(value="android", link="lept"),
     @Platform(value="windows", link="liblept", preload={"libwinpthread-1", "libgcc_s_sjlj-1", "libgcc_s_seh-1", "libstdc++-6",
         "zlib1", "libgif-6", "libjpeg-62", "libpng16-16", "libtiff-5", "libwebp-5", "liblept-4"}),
-    @Platform(value="windows-x86", preloadpath={"C:/Program Files (x86)/mingw-w64/i686-4.9.1-posix-sjlj-rt_v3-rev3/mingw32/bin/"}),
-    @Platform(value="windows-x86_64", preloadpath={"C:/Program Files/mingw-w64/x86_64-4.9.1-posix-seh-rt_v3-rev3/mingw64/bin/"}) })
+    @Platform(value="windows-x86", preloadpath="C:/Program Files (x86)/mingw-w64/i686-4.9.1-posix-sjlj-rt_v3-rev3/mingw32/bin/"),
+    @Platform(value="windows-x86_64", preloadpath="C:/Program Files/mingw-w64/x86_64-4.9.1-posix-seh-rt_v3-rev3/mingw64/bin/") })
 public class lept implements InfoMapper {
     public void map(InfoMap infoMap) {
         infoMap.put(new Info("LEPT_DLL", "L_END_LIST").cppTypes().annotations())